feat/choose-users: split slack types from model, scaffold choose command
