+++ package.json
@@ -22,6 +22,7 @@
     "@slack/bolt": "^3.12.2",
     "@types/node": "^18.11.18",
     "dotenv": "^16.0.3",
+    "lodash": "^4.17.21",
     "serverless": "^3.26.0",
     "typescript": "^4.9.4",
     "winston": "^3.8.2",
