+++ src/types/index.ts
@@ -1,14 +1,3 @@
 export * from "./model";
-
-export interface ICommand {
-  input?: string;
-  handle(): any;
-}
-
-export interface Dictionary<T> {
-  [Key: string]: T;
-}
-
-export class StringDictionary {
-  Entries: Dictionary<string> = {};
-}
+export * from "./slack";
+export * from "./utils";
