+++ src/types/utils.ts
@@ -1,0 +1,12 @@
+export interface ICommand {
+  input?: string;
+  handle(): any;
+}
+
+export interface Dictionary<T> {
+  [Key: string]: T;
+}
+
+export class StringDictionary {
+  Entries: Dictionary<string> = {};
+}
